Ran clippy and fixed warnings
