+++ src/cpu/address_modes.rs
@@ -1,4 +1,4 @@
-use crate::cpu::cpu;
+use super::Cpu6502;
 
 #[derive(Debug)]
 pub enum AddressMode {
@@ -19,7 +19,7 @@ pub enum AddressMode {
 /// Address mode: Implicit
 /// No extra data is needed
 /// Need to target accumulator for instructions like PHA
-pub fn imp(cpu6502: &mut cpu::Cpu6502) -> u8 {
+pub fn imp(cpu6502: &mut Cpu6502) -> u8 {
     cpu6502.fetched_data = cpu6502.accumulator;
     0
 }
@@ -27,7 +27,7 @@ pub fn imp(cpu6502: &mut cpu::Cpu6502) -> u8 {
 /// Address mode: Immediate
 /// The instruction expected the next byte to be used as data
 /// Increment the program counter to access that
-pub fn imm(cpu6502: &mut cpu::Cpu6502) -> u8 {
+pub fn imm(cpu6502: &mut Cpu6502) -> u8 {
     cpu6502.addr_abs = cpu6502.program_counter;
     cpu6502.program_counter = cpu6502.program_counter.wrapping_add(1);
     0
@@ -35,38 +35,38 @@ pub fn imm(cpu6502: &mut cpu::Cpu6502) -> u8 {
 
 /// Address mode: Zero Page
 /// Allows you to absolutely access the first 256 bytes of a location
-pub fn zp0(cpu6502: &mut cpu::Cpu6502) -> u8 {
+pub fn zp0(cpu6502: &mut Cpu6502) -> u8 {
     let address = cpu6502.read(cpu6502.program_counter);
     cpu6502.program_counter = cpu6502.program_counter.wrapping_add(1);
-    cpu6502.addr_abs = (address & 0x00FF) as u16;
+    cpu6502.addr_abs = address as u16;
     0
 }
 
 /// Address mode: Zero Page with X Offset
 /// Same as zero page but with the X address added
-pub fn zpx(cpu6502: &mut cpu::Cpu6502) -> u8 {
+pub fn zpx(cpu6502: &mut Cpu6502) -> u8 {
     let address = cpu6502.read(cpu6502.program_counter).wrapping_add(cpu6502.x_register);
     cpu6502.program_counter = cpu6502.program_counter.wrapping_add(1);
-    cpu6502.addr_abs = (address & 0x00FF) as u16;
+    cpu6502.addr_abs = address as u16;
     0
 }
 
 /// Address mode: Zero Page with Y Offset
 /// Same as zero page but with Y addressadded
-pub fn zpy(cpu6502: &mut cpu::Cpu6502) -> u8 {
+pub fn zpy(cpu6502: &mut Cpu6502) -> u8 {
     let address = cpu6502.read(cpu6502.program_counter).wrapping_add(cpu6502.y_register);
     cpu6502.program_counter = cpu6502.program_counter.wrapping_add(1);
-    cpu6502.addr_abs = (address & 0x00FF) as u16;
+    cpu6502.addr_abs = address as u16;
     0
 }
 
 /// Address mode: Relative
 /// Branching instructions can't jump any further than 127 memory locations
-pub fn rel(cpu6502: &mut cpu::Cpu6502) -> u8 {
+pub fn rel(cpu6502: &mut Cpu6502) -> u8 {
     cpu6502.addr_rel = cpu6502.read(cpu6502.program_counter) as u16;
     cpu6502.program_counter = cpu6502.program_counter.wrapping_add(1);
     if cpu6502.addr_rel & 0x80 != 0 {
-        cpu6502.addr_rel = cpu6502.addr_rel | 0xFF00;
+        cpu6502.addr_rel |= 0xFF00;
     }
 
     0
@@ -74,7 +74,7 @@ pub fn rel(cpu6502: &mut cpu::Cpu6502) -> u8 {
 
 /// Address mode: Asbolute
 /// Read the low and high from the next two instructions
-pub fn abs(cpu6502: &mut cpu::Cpu6502) -> u8 {
+pub fn abs(cpu6502: &mut Cpu6502) -> u8 {
     cpu6502.addr_abs = get_absolute_address(cpu6502);
     0
 }
@@ -82,7 +82,7 @@ pub fn abs(cpu6502: &mut cpu::Cpu6502) -> u8 {
 /// Address mode: Absolute with X offset
 /// Read the low and high from the next two instructions.
 /// Then adds the X register to the result
-pub fn abx(cpu6502: &mut cpu::Cpu6502) -> u8 {
+pub fn abx(cpu6502: &mut Cpu6502) -> u8 {
     let address = get_absolute_address(cpu6502);
     cpu6502.addr_abs = address.wrapping_add(cpu6502.x_register as u16);
 
@@ -96,7 +96,7 @@ pub fn abx(cpu6502: &mut cpu::Cpu6502) -> u8 {
 /// Address mode: Absolute with X offset
 /// Read the low and high from the next two instructions.
 /// Then adds the Y register to the result
-pub fn aby(cpu6502: &mut cpu::Cpu6502) -> u8 {
+pub fn aby(cpu6502: &mut Cpu6502) -> u8 {
     let address = get_absolute_address(cpu6502);
     cpu6502.addr_abs = address.wrapping_add(cpu6502.y_register as u16);
 
@@ -110,7 +110,7 @@ pub fn aby(cpu6502: &mut cpu::Cpu6502) -> u8 {
 /// Address mode: Indirect
 /// There is a bug in the 6502 implementation where if the lower byte is equal to 0xFF it then needs to cross
 /// a page boundary. It does this incorrectly and instead wraps around the same page so we need to replicate that.
-pub fn ind(cpu6502: &mut cpu::Cpu6502) -> u8 {
+pub fn ind(cpu6502: &mut Cpu6502) -> u8 {
     let pointer_address = get_absolute_address(cpu6502);
 
     if (pointer_address & 0x00FF) == 0x00FF {
@@ -129,12 +129,12 @@ pub fn ind(cpu6502: &mut cpu::Cpu6502) -> u8 {
 /// Address mode: Indirext X
 /// The supplied address is offset by X to index a location in page 0x00
 /// The actual address is then read from this location
-pub fn izx(cpu6502: &mut cpu::Cpu6502) -> u8 {
+pub fn izx(cpu6502: &mut Cpu6502) -> u8 {
     let address = cpu6502.read(cpu6502.program_counter);
     cpu6502.program_counter = cpu6502.program_counter.wrapping_add(1);
 
-    let low_address = (address.wrapping_add(cpu6502.x_register)) & 0x00FF;
-    let high_address = (address.wrapping_add(cpu6502.x_register + 1)) & 0x00FF;
+    let low_address = address.wrapping_add(cpu6502.x_register);
+    let high_address = address.wrapping_add(cpu6502.x_register + 1);
     let low = cpu6502.read(low_address as u16) as u16;
     let high = cpu6502.read(high_address as u16) as u16;
 
@@ -147,7 +147,7 @@ pub fn izx(cpu6502: &mut cpu::Cpu6502) -> u8 {
 /// The supplied address is a location in page 0x00
 /// The address is then read from this and then offset by Y
 /// If a page boundary occurs, an additional clock cycle is required
-pub fn izy(cpu6502: &mut cpu::Cpu6502) -> u8 {
+pub fn izy(cpu6502: &mut Cpu6502) -> u8 {
     let pointer_address = cpu6502.read(cpu6502.program_counter) as u16;
     cpu6502.program_counter = cpu6502.program_counter.wrapping_add(1);
 
@@ -161,12 +161,11 @@ pub fn izy(cpu6502: &mut cpu::Cpu6502) -> u8 {
     0
 }
 
-fn get_absolute_address(cpu6502: &mut cpu::Cpu6502) -> u16 {
+fn get_absolute_address(cpu6502: &mut Cpu6502) -> u16 {
     let low = cpu6502.read(cpu6502.program_counter) as u16;
     cpu6502.program_counter = cpu6502.program_counter.wrapping_add(1);
     let high = cpu6502.read(cpu6502.program_counter) as u16;
     cpu6502.program_counter = cpu6502.program_counter.wrapping_add(1);
 
-    let address = (high << 8) | low;
-    return address;
+    (high << 8) | low
 }
