+++ src/cpu/opcode_table.rs
@@ -1,8 +1,10 @@
-use crate::cpu::cpu;
-use crate::cpu::address_modes;
-use crate::cpu::opcodes;
+use super::Cpu6502;
+use super::address_modes;
+use super::opcodes;
 
-pub const OPCODE_TABLE: [(&str, fn(&mut cpu::Cpu6502) -> u8, fn(&mut cpu::Cpu6502) -> u8, address_modes::AddressMode, u8); 256] = [
+type Opcode<'a> = (&'a str, fn(&mut Cpu6502) -> u8, fn(&mut Cpu6502) -> u8, address_modes::AddressMode, u8);
+
+pub const OPCODE_TABLE: [Opcode; 256] = [
     ( "brk", opcodes::brk, address_modes::imm, address_modes::AddressMode::Imm, 7 ),( "ora", opcodes::ora, address_modes::izx, address_modes::AddressMode::Izx, 6 ),( "???", opcodes::xxx, address_modes::imp, address_modes::AddressMode::Imp, 2 ),( "???", opcodes::xxx, address_modes::imp, address_modes::AddressMode::Imp, 8 ),( "???", opcodes::nop, address_modes::imp, address_modes::AddressMode::Imp, 3 ),( "ora", opcodes::ora, address_modes::zp0, address_modes::AddressMode::Zp0, 3 ),( "asl", opcodes::asl, address_modes::zp0, address_modes::AddressMode::Zp0, 5 ),( "???", opcodes::xxx, address_modes::imp, address_modes::AddressMode::Imp, 5 ),( "php", opcodes::php, address_modes::imp, address_modes::AddressMode::Imp, 3 ),( "ora", opcodes::ora, address_modes::imm, address_modes::AddressMode::Imm, 2 ),( "asl", opcodes::asl, address_modes::imp, address_modes::AddressMode::Imp, 2 ),( "???", opcodes::xxx, address_modes::imp, address_modes::AddressMode::Imp, 2 ),( "???", opcodes::nop, address_modes::imp, address_modes::AddressMode::Imp, 4 ),( "ora", opcodes::ora, address_modes::abs, address_modes::AddressMode::Abs, 4 ),( "asl", opcodes::asl, address_modes::abs, address_modes::AddressMode::Abs, 6 ),( "???", opcodes::xxx, address_modes::imp, address_modes::AddressMode::Imp, 6 ),
     ( "bpl", opcodes::bpl, address_modes::rel, address_modes::AddressMode::Rel, 2 ),( "ora", opcodes::ora, address_modes::izy, address_modes::AddressMode::Izy, 5 ),( "???", opcodes::xxx, address_modes::imp, address_modes::AddressMode::Imp, 2 ),( "???", opcodes::xxx, address_modes::imp, address_modes::AddressMode::Imp, 8 ),( "???", opcodes::nop, address_modes::imp, address_modes::AddressMode::Imp, 4 ),( "ora", opcodes::ora, address_modes::zpx, address_modes::AddressMode::Zpx, 4 ),( "asl", opcodes::asl, address_modes::zpx, address_modes::AddressMode::Zpx, 6 ),( "???", opcodes::xxx, address_modes::imp, address_modes::AddressMode::Imp, 6 ),( "clc", opcodes::clc, address_modes::imp, address_modes::AddressMode::Imp, 2 ),( "ora", opcodes::ora, address_modes::aby, address_modes::AddressMode::Aby, 4 ),( "???", opcodes::nop, address_modes::imp, address_modes::AddressMode::Imp, 2 ),( "???", opcodes::xxx, address_modes::imp, address_modes::AddressMode::Imp, 7 ),( "???", opcodes::nop, address_modes::imp, address_modes::AddressMode::Imp, 4 ),( "ora", opcodes::ora, address_modes::abx, address_modes::AddressMode::Abx, 4 ),( "asl", opcodes::asl, address_modes::abx, address_modes::AddressMode::Abx, 7 ),( "???", opcodes::xxx, address_modes::imp, address_modes::AddressMode::Imp, 7 ),
     ( "jsr", opcodes::jsr, address_modes::abs, address_modes::AddressMode::Abs, 6 ),( "and", opcodes::and, address_modes::izx, address_modes::AddressMode::Izx, 6 ),( "???", opcodes::xxx, address_modes::imp, address_modes::AddressMode::Imp, 2 ),( "???", opcodes::xxx, address_modes::imp, address_modes::AddressMode::Imp, 8 ),( "bit", opcodes::bit, address_modes::zp0, address_modes::AddressMode::Zp0, 3 ),( "and", opcodes::and, address_modes::zp0, address_modes::AddressMode::Zp0, 3 ),( "rol", opcodes::rol, address_modes::zp0, address_modes::AddressMode::Zp0, 5 ),( "???", opcodes::xxx, address_modes::imp, address_modes::AddressMode::Imp, 5 ),( "plp", opcodes::plp, address_modes::imp, address_modes::AddressMode::Imp, 4 ),( "and", opcodes::and, address_modes::imm, address_modes::AddressMode::Imm, 2 ),( "rol", opcodes::rol, address_modes::imp, address_modes::AddressMode::Imp, 2 ),( "???", opcodes::xxx, address_modes::imp, address_modes::AddressMode::Imp, 2 ),( "bit", opcodes::bit, address_modes::abs, address_modes::AddressMode::Abs, 4 ),( "and", opcodes::and, address_modes::abs, address_modes::AddressMode::Abs, 4 ),( "rol", opcodes::rol, address_modes::abs, address_modes::AddressMode::Abs, 6 ),( "???", opcodes::xxx, address_modes::imp, address_modes::AddressMode::Imp, 6 ),
